Consolidate editor code
